Several snags handled including poll cell size, forum comment controller, info tab image, and master account function implementations (really should have used more than one branch but whatever)

+++ Greek Life/Greek Life/CommentViewController.swift
@@ -29,8 +29,6 @@ class CommentTableViewCell: UITableViewCell {
     }
     
 }
-
-
 class CommentViewController: UIViewController, UITableViewDelegate, UITableViewDataSource {
     var rowHeight:CGFloat = 0
     var ref: DatabaseReference!
@@ -38,40 +36,43 @@ class CommentViewController: UIViewController, UITableViewDelegate, UITableViewD
     let user = LoggedIn.User["Username"] as! String
     let userName = "\(LoggedIn.User["First Name"] as! String) \(LoggedIn.User["Last Name"] as! String)"
     let userId = LoggedIn.User["UserID"] as! String
-
+    
+    var InteractedCommentIndex = 0
+    
+    @IBOutlet weak var TableView: UITableView!
+    @IBOutlet weak var LeaveComment: UIButton!
+    @IBOutlet weak var CommentBox: UITextView!
+    
     @IBAction func BackToForum(_ sender: Any) {
         self.dismiss(animated: true, completion: nil)
     }
     
-    //---Prototype cell component declarations---//
-    @IBOutlet weak var TableView: UITableView!
-    
-    //---Function to add a comment to a post---//
-    @IBOutlet weak var CommentBox: UITextField!
-    @IBOutlet weak var LeaveComment: UIBarButtonItem! //button
+    //--Process comment --//
     @IBAction func LeaveComment(_ sender: Any) {
         let date = Date().timeIntervalSince1970
         let comment = CommentBox.text
         let components = comment!.components(separatedBy: .whitespacesAndNewlines)
         let PostWords = components.filter { !$0.isEmpty }
         if comment == "" {
-            let emptyError = UIAlertController(title: "Empty Comment", message: "Your Comment cannot be empty", preferredStyle: UIAlertControllerStyle.alert)
+            let emptyError = UIAlertController(title: "Empty Comment", message: "Your comment cannot be empty", preferredStyle: UIAlertControllerStyle.alert)
             let okAction = UIAlertAction(title: "Ok", style: UIAlertActionStyle.default)
             emptyError.addAction(okAction)
             self.present(emptyError, animated: true, completion: nil)
         }
         else if PostWords.count > 200 {
-            let emptyError = UIAlertController(title: "Too Large", message: "Your Comment cannot be larger than 200 words!", preferredStyle: UIAlertControllerStyle.alert)
+            let emptyError = UIAlertController(title: "Too Large", message: "Your comment cannot be larger than 200 words!", preferredStyle: UIAlertControllerStyle.alert)
             let okAction = UIAlertAction(title: "Ok", style: UIAlertActionStyle.default)
             emptyError.addAction(okAction)
             self.present(emptyError, animated: true, completion: nil)
         }
         else {
+            let postId = UUID().uuidString
             let NewComment = [
                 "Post": comment!,
                 "Poster": self.userName,
                 "Epoch": date,
-                "PosterId": userId
+                "PosterId": userId,
+                "CommentId": postId
                 ] as [String : Any]
             PostData(newPostData: NewComment, completion: { (success, error) in
                 self.CommentBox.text = ""
@@ -80,25 +81,24 @@ class CommentViewController: UIViewController, UITableViewDelegate, UITableViewD
         }
     }
     
-    //---Write comment data to database---//
-    func PostData(newPostData: Dictionary<String, Any>, completion: @escaping (Bool, Error?) -> Void){
-        ref = Database.database().reference()
-        let commentID = UUID().uuidString
-        self.ref.child("Forum").child(Postings.AllPosts![Postings.myIndex].PostId).child("Comments").child(commentID).setValue(newPostData)
-        completion(true, nil)
-    }
     override func viewDidLoad() {
         super.viewDidLoad()
         ReadCommentsForPost()
         self.TableView.allowsSelection = false
-        
-
-        //---Add styles---//
-            }
+        self.CommentBox.layer.cornerRadius = 10
+    }
 
     override func didReceiveMemoryWarning() {
         super.didReceiveMemoryWarning()
         // Dispose of any resources that can be recreated.
+    }
+    
+    //---Write comment data to database---//
+    func PostData(newPostData: Dictionary<String, Any>, completion: @escaping (Bool, Error?) -> Void){
+        ref = Database.database().reference()
+        let commentId = newPostData["CommentId"] as! String
+        self.ref.child("Forum").child(Postings.AllPosts![Postings.myIndex].PostId).child("Comments").child(commentId).setValue(newPostData)
+        completion(true, nil)
     }
     
     public func ReadCommentsForPost() {
@@ -113,9 +113,11 @@ class CommentViewController: UIViewController, UITableViewDelegate, UITableViewD
                         if let commenter = postDictionary["Poster"] as? String {
                             if let commentDate = postDictionary["Epoch"] as? Double {
                                 if let commenterId = postDictionary["PosterId"] as? String {
+                                    if let commentID = postDictionary["CommentId"] as? String {
                             let pDate = CreateDate.getCurrentDate(epoch: commentDate)
-                                let aComment = Comment(Poster: commenter, PostDate: pDate, PostEpoch: commentDate, Post: comment, PosterId: commenterId)
+                                    let aComment = Comment(Poster: commenter, PostDate: pDate, PostEpoch: commentDate, Post: comment, PosterId: commenterId, CommentId: commentID)
                                 self.CommentList.append(aComment)
+                            }
                         }
                       }
                     }
@@ -132,63 +134,73 @@ class CommentViewController: UIViewController, UITableViewDelegate, UITableViewD
     }
     
     @objc func DeleteBtn(button: UIButton) {
-        //delete it
+        let cell = button.superview?.superviewOfClassType(UITableViewCell.self) as! UITableViewCell
+        let tbl = cell.superviewOfClassType(UITableView.self) as! UITableView
+        let indexPath = tbl.indexPath(for: cell)
+        self.InteractedCommentIndex = indexPath!.row
+        let delete = UIAlertController(title: "Delete", message: "Are you sure you would like to delete your comment?", preferredStyle: UIAlertControllerStyle.alert)
+        let okAction = UIAlertAction(title: "Delete", style: UIAlertActionStyle.default, handler: DeleteComment)
+        let cancelAction = UIAlertAction(title: "Cancel", style: UIAlertActionStyle.default)
+        delete.addAction(cancelAction)
+        delete.addAction(okAction)
+        self.present(delete, animated: true, completion: nil)
+    }
+    
+    func DeleteComment(action: UIAlertAction) {
+        FirebaseDatabase.Database.database().reference().child("Forum").child(Postings.AllPosts![Postings.myIndex].PostId).child("Comments").child(CommentList[self.InteractedCommentIndex - 1].CommentId).removeValue()
     }
     
     public func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: "CommentCell", for: indexPath) as! CommentTableViewCell
         if(indexPath.row == 0){
-            cell.layer.borderWidth = 1
             cell.Comment.text = Postings.AllPosts![Postings.myIndex].Post
             cell.CommenterName.text = Postings.AllPosts![Postings.myIndex].Poster
-            cell.CommenterName.textColor = UIColor.purple
             let timeSince = CreateDate.getTimeSince(epoch: Postings.AllPosts![Postings.myIndex].PostDate) //4 days
             cell.CommentDate.text = timeSince
-            
             GenericTools.FrameToFitTextView(View: cell.Comment)
-            self.rowHeight = cell.Comment.frame.origin.y + cell.Comment.frame.size.height
+            cell.layer.borderWidth = 1
+            self.rowHeight = cell.Comment.frame.origin.y + cell.Comment.frame.size.height + 30
             cell.Delete.isHidden = true
 
             return cell
         }
-        if CommentList[indexPath.row - 1].PosterId != self.userId {
-            cell.Delete.isHidden = true
-        }
-        else {
-            cell.Delete.addTarget(self, action: #selector(DeleteBtn(button:)), for: .touchUpInside)
-        }
-
+        cell.Delete.isHidden = false
         //--Set Content--//
         if CommentList.count > 0 {
+            cell.backgroundColor = .white
+            if indexPath.row % 2 == 0 {
+                cell.backgroundColor = UIColor(displayP3Red: 250/255, green: 234/255, blue: 234/255, alpha: 0.5)
+            }
+            else {
+            }
         cell.layer.borderWidth = 0.1
-        cell.Comment.text = CommentList[indexPath.row - 1].Post
         cell.CommenterName.text = CommentList[indexPath.row - 1].Poster
-        cell.CommenterName.textColor = .blue
-        if cell.CommenterName.text == self.user {
-            cell.CommenterName.textColor = .green
-        }
-        let timeSince = CreateDate.getTimeSince(epoch: CommentList[indexPath.row - 1].PostEpoch) //<4 days
+        let timeSince = CreateDate.getTimeSince(epoch: CommentList[indexPath.row - 1].PostEpoch)
         cell.CommentDate.text = timeSince
-        
-        //--Change cell height--//
-        let oldHeight = cell.Comment.frame.size.height
-        GenericTools.FrameToFitTextView(View: cell.Comment)
-        let newHeight = cell.Comment.frame.size.height
-        let cellHeight = newHeight + 20 + 12
-        let heightDifference = oldHeight - newHeight
-        cell.CommentDate.frame.size.height -= heightDifference
-        self.rowHeight = cellHeight
-        
+        cell.Comment.text = CommentList[indexPath.row - 1].Post
+        if CommentList[indexPath.row - 1].PosterId  == self.userId {
+            cell.CommenterName.textColor = .blue
         }
-        return(cell)
+        //--Change cell height--//
+        GenericTools.FrameToFitTextView(View: cell.Comment)
+        cell.CommenterName.frame.origin.y = 20
+        cell.CommentDate.frame.origin.y = cell.CommenterName.frame.origin.y
+        cell.Comment.frame.origin.y = cell.CommenterName.frame.origin.y + cell.CommenterName.frame.size.height
+        cell.Delete.frame.origin.y = (cell.Comment.frame.origin.y + cell.Comment.frame.size.height) + 5
+            if CommentList[indexPath.row - 1].PosterId != self.userId && self.user != "Master" {
+                cell.Delete.isHidden = true
+                self.rowHeight = cell.Comment.frame.origin.y + cell.Comment.frame.size.height
+            }
+            else {
+                cell.Delete.addTarget(self, action: #selector(DeleteBtn(button:)), for: .touchUpInside)
+                self.rowHeight = cell.Delete.frame.origin.y + cell.Delete.frame.size.height + 5
+            }
+        }
+        return cell
     }
     
     public func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        var postings = CommentList.count
-        if postings == 0 {
-            postings = 1
-        }
-        return postings + 1
+        return CommentList.count + 1
     }
     
     public func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
